requirements.txt
Appium_Python_Client==2.1.2
mysql_connector_repackaged==0.3.1
pandas==1.3.5
requests==2.27.1
